fix #791

+++ backend/src/services/access.service.ts
@@ -6,6 +6,7 @@ import {
     FindOptionsWhere,
     ILike,
     InsertResult,
+    Not,
     Repository,
 } from 'typeorm';
 import AccessGroup from '@common/entities/auth/accessgroup.entity';
@@ -60,7 +61,10 @@ export class AccessService {
             createdAt: rawAccessGroup.createdAt,
             creator: rawAccessGroup.creator?.userDto ?? null,
             hidden: rawAccessGroup.hidden,
-            memberships: [],
+            memberships:
+                rawAccessGroup.memberships?.map(
+                    (membership) => membership.groupMembershipDto,
+                ) ?? [],
             name: rawAccessGroup.name,
             type: rawAccessGroup.type,
             updatedAt: rawAccessGroup.updatedAt,
@@ -82,6 +86,7 @@ export class AccessService {
             memberships: [
                 {
                     user: { uuid: user.uuid },
+                    canEditGroup: true, // the creator can always edit the group
                 },
             ],
             creator: user,
@@ -250,6 +255,21 @@ export class AccessService {
         accessGroupUUID: string,
         userUUID: string,
     ): Promise<AccessGroup> {
+        const usersWithEditRights = await this.groupMembershipRepository.count({
+            where: {
+                accessGroup: { uuid: accessGroupUUID },
+                // where user is NOT matching userUUID
+                user: { uuid: Not(userUUID) },
+                canEditGroup: true,
+            },
+        });
+
+        if (usersWithEditRights === 0) {
+            throw new ConflictException(
+                'Cannot remove the last user with edit rights',
+            );
+        }
+
         await this.groupMembershipRepository.delete({
             accessGroup: { uuid: accessGroupUUID },
             user: { uuid: userUUID },
@@ -470,16 +490,16 @@ export class AccessService {
 
     async setExpireDate(
         uuid: string,
-        expireDate: Date | null,
+        expireDate: Date | 'never',
     ): Promise<GroupMembershipDto> {
         const agu = await this.groupMembershipRepository.findOneOrFail({
-            where: { uuid },
+            where: {
+                accessGroup: { uuid },
+            },
         });
-        if (expireDate === null) {
-            delete agu.expirationDate;
-        } else {
-            agu.expirationDate = expireDate;
-        }
+        // TODO: check how to properly set the expireDate to null in a type-safe way
+        // @ts-ignore
+        agu.expirationDate = expireDate === 'never' ? null : expireDate;
         return (await this.groupMembershipRepository.save(
             agu,
         )) as unknown as GroupMembershipDto;
